Improve testing
* Add support for powerpc7 Altivec (WIP)

Some functions still do not work:
 - load2/3/4 and store2/3/4
 - reinterptret cast between f64 and i64
 - conversion between f16 u16 type.

Addv function is now usable with any basics types (including integer
types), but modification still need to be made for it to work with arm and x86.

* Degree>=2 load/store for ppc architecture (64 bits types)

* Correct cast between i64 and f64 (emulated).

* Correctly convert u16 to and from f16

* Load of degree 2, 3 and 4 for every types but f16.

* Improve tests for load and store functions

 - Test that loadn correctly ravels data (ie: {0, 1, 0, 1} in load2
 results in {0, 0}, {1, 1}).
 - Test that loadu and storeu function correctly load and store to
 unaligned adresses by forcing the use of unaligned vectors in the
 tests.

* Loadx and storex for ppc architecture

* Don't define addv for non float types

* Corrects bug in gt for sse2 and ne for avx

Corrects:
 - a bug in gt for sse, that was present when comparing hugely
positive number with hugely negative number.
 - a bug that affected avx ne, when comparing NaN with other numbers.

* Test now generate random number over the whole range of int/float

* Correct comparison in x86 sse2

* Ensure all test pass for power

* Correct abs function for avx2

* Correct bug in shr for neon128

* Correct fma, fms, fmna and fmns tests for NEON128

* Remove useless exception for comparison on intel

* Correct ne for AVX512 for NaN

* Improve random number generation for f16

* Add zip and unzip

* Merge + but ppc support for now (merge only to get the test improvment in master)

* Add (generated) documentation file to the .gitignore

* Improve f16 to f32 and f32 to f16 conversion

Fully support NaN/Inf, denormals...

* Correct warnings

+++ egg/modules/fixed_point/gen_tests.py
@@ -74,7 +74,7 @@ check = """\
 limits = """\
 template <uint8_t lf, uint8_t rt>
 static double __get_numeric_precision() {
-  return ldexpf(1.0, -(int)rt);
+  return (double)ldexpf(1.0, -(int)rt);
 }
 
 """
@@ -135,8 +135,8 @@ int main() {{
   for (size_t i = 0; i < v_size; i++) {{
     tab0_fp[i] = __gen_random_val<{lf}, {rt}>();
     tab1_fp[i] = __gen_random_val<{lf}, {rt}>();
-    tab0_f[i] = float(tab0_fp[i]);
-    tab1_f[i] = float(tab1_fp[i]);
+    tab0_f[i] = double(tab0_fp[i]);
+    tab1_f[i] = double(tab1_fp[i]);
   }}
 
   vec_t v0_fp = nsimd::fixed_point::loadu<vec_t>(tab0_fp);
@@ -260,9 +260,9 @@ int main() {{
     tab0_fp[i] = __gen_random_val<{lf}, {rt}>();
     tab1_fp[i] = __gen_random_val<{lf}, {rt}>();
     tab2_fp[i] = __gen_random_val<{lf}, {rt}>();
-    tab0_f[i] = float(tab0_fp[i]);
-    tab1_f[i] = float(tab1_fp[i]);
-    tab2_f[i] = float(tab2_fp[i]);
+    tab0_f[i] = double(tab0_fp[i]);
+    tab1_f[i] = double(tab1_fp[i]);
+    tab2_f[i] = double(tab2_fp[i]);
   }}
 
   vec_t v0_fp = nsimd::fixed_point::loadu<vec_t>(tab0_fp);
@@ -328,7 +328,7 @@ int main() {{
 
   for (size_t i = 0; i < v_size; i++) {{
     tab0_fp[i] = __gen_random_val<{lf}, {rt}>();
-    tab0_f[i] = float(tab0_fp[i]);
+    tab0_f[i] = double(tab0_fp[i]);
   }}
 
   vec_t v0_fp = nsimd::fixed_point::loadu<vec_t>(tab0_fp);
